[refactor] orderItem 디티오 수정
- 오더아이템은 따로 Api를 보내는게 아니기 때문에 리스폰스 제외하고 리퀘스트 디티오만 두기로 결정
