+++ src/main/java/com/p1/nomnom/orderItems/dto/OrderItemDto.java
@@ -1,4 +1,4 @@
-package com.p1.nomnom.orderItems.dto.request;
+package com.p1.nomnom.orderItems.dto;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
@@ -10,7 +10,7 @@ import java.util.UUID;
 @Getter
 @NoArgsConstructor(access = lombok.AccessLevel.PRIVATE)
 @AllArgsConstructor(access = lombok.AccessLevel.PRIVATE)
-public class OrderItemRequestDto {
+public class OrderItemDto {
 
     @NotNull
     private UUID foodId;
